requirements-dev.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0

cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.12.2
idna==2.10
isodate==0.6.0
lxml==4.6.2
Markdown==3.3.3
Pillow==8.1.0
pycryptodome==3.10.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.2
zeep==4.0.0
